+++ static/types/entities/index.d.ts
@@ -1,0 +1,6 @@
+import "./DefaultEntities";
+import "./Resources";
+import "./Subsystems";
+import "./Things";
+import "./ThingTemplates";
+import "./ThingShapes";