+++ ui/index.d.ts
@@ -1,0 +1,5 @@
+import "./types/UIBaseTypes"
+import "./types/UIBaseWidgets"
+import "./types/UIDefaultWidgets"
+import "../static/types/TWBaseTypes"
+import "../static/types/entities"